the_givers/requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
newsapi-python==0.2.6
numpy==1.19.4
pandas==1.1.4
pathspec==0.8.0
plotly==4.12.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
regex==2020.9.27
requests==2.24.0
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.0.33
Werkzeug==1.0.1
WTForms==2.3.3